fix: Rename "monkeypatch" modules to "snekpatch"
