+++ snekoil/unsnekpatch.py
@@ -8,6 +8,6 @@ from snekoil.time import _original_sleep
 builtins.range = _original_range
 time.sleep = _original_sleep
 
-# Remove the monkeypatch module from sys.modules so that "import snekoil.monkeypatch" works again
-if 'snekoil.monkeypatch' in sys.modules:
-    del sys.modules['snekoil.monkeypatch']
+# Remove the snekpatch module from sys.modules so that "import snekoil.snekpatch" works again
+if 'snekoil.snekpatch' in sys.modules:
+    del sys.modules['snekoil.snekpatch']
